feat:modifiy point from Integer to Float

+++ src/main/java/com/iko/iko/controller/member/dto/response/MemberResponseDto.java
@@ -29,7 +29,7 @@ public class MemberResponseDto {
 
     private String birthday;
 
-    private Integer point;
+    private Float point;
 
     private Date createdAt;
